refactor: moved transaction preview component to comon components
