+++ client/src/components/TransactionPreview/components/TPTable/TPTable.types.ts
@@ -1,4 +1,4 @@
-import { Transaction } from '../../../../../../types/Transaction';
+import { Transaction } from '../../../../types/Transaction'
 
 export interface IProps {
 	categoryId: number
